requirements.txt
amqp==2.5.0
Babel==2.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.0.13
django-appconf==1.0.3
django-compressor==2.3
django-environ==0.4.5
django-extensions==2.2.1
django-extra-views==0.11.0
django-guardian==2.0.0
django-haystack==2.8.1
django-localflavor==2.2
django-oscar==1.6.1
django-oscar-api==1.6.1
django-oscar-paymentexpress==0.1.1
django-oscar-paypal==1.0.0
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
djangorestframework==3.10.1
factory-boy==2.12.0
Faker==2.0.0
flower==0.9.3
idna==2.8
kombu==4.6.3
mock==2.0.0
pbr==5.4.1
phonenumbers==8.10.15
Pillow==6.1.0
purl==1.5
pycountry==19.7.15
python-dateutil==2.8.0
python-stdnum==1.11
pytz==2019.1
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sorl-thumbnail==12.4.1
sqlparse==0.3.0
text-unidecode==1.2
tornado==5.1.1
Unidecode==1.0.23
urllib3==1.25.3
vine==1.3.0
